+++ exporters/otlp/otlplog/otlploghttp/internal/transform/log.go
@@ -12,9 +12,9 @@ import (
 	"math"
 	"time"
 
-	cpb "go.opentelemetry.io/proto/otlphttp/common/v1"
-	lpb "go.opentelemetry.io/proto/otlphttp/logs/v1"
-	rpb "go.opentelemetry.io/proto/otlphttp/resource/v1"
+	cpb "go.opentelemetry.io/proto/slim/otlp/common/v1"
+	lpb "go.opentelemetry.io/proto/slim/otlp/logs/v1"
+	rpb "go.opentelemetry.io/proto/slim/otlp/resource/v1"
 
 	"go.opentelemetry.io/otel/attribute"
 	api "go.opentelemetry.io/otel/log"