+++ exporters/otlp/otlpmetric/otlpmetrichttp/client_test.go
@@ -21,8 +21,8 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	colmetricpb "go.opentelemetry.io/proto/otlphttp/collector/metrics/v1"
-	mpb "go.opentelemetry.io/proto/otlphttp/metrics/v1"
+	colmetricpb "go.opentelemetry.io/proto/slim/otlp/collector/metrics/v1"
+	mpb "go.opentelemetry.io/proto/slim/otlp/metrics/v1"
 
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"