+++ exporters/otlp/otlpmetric/otlpmetrichttp/internal/transform/metricdata_test.go
@@ -13,9 +13,9 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	cpb "go.opentelemetry.io/proto/otlphttp/common/v1"
-	mpb "go.opentelemetry.io/proto/otlphttp/metrics/v1"
-	rpb "go.opentelemetry.io/proto/otlphttp/resource/v1"
+	cpb "go.opentelemetry.io/proto/slim/otlp/common/v1"
+	mpb "go.opentelemetry.io/proto/slim/otlp/metrics/v1"
+	rpb "go.opentelemetry.io/proto/slim/otlp/resource/v1"
 
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/sdk/instrumentation"