+++ exporters/otlp/otlptrace/otlptracehttp/v2/client.go
@@ -18,8 +18,8 @@ import (
 	"sync"
 	"time"
 
-	coltracepb "go.opentelemetry.io/proto/otlphttp/collector/trace/v1"
-	tracepb "go.opentelemetry.io/proto/otlphttp/trace/v1"
+	coltracepb "go.opentelemetry.io/proto/slim/otlp/collector/trace/v1"
+	tracepb "go.opentelemetry.io/proto/slim/otlp/trace/v1"
 	"google.golang.org/protobuf/proto"
 
 	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracehttp/v2/internal"