+++ exporters/otlp/otlptrace/otlptracehttp/v2/internal/otlpjson/export_trace_service_request.go
@@ -19,10 +19,10 @@ import (
 	"strconv"
 	"strings"
 
-	coltracepb "go.opentelemetry.io/proto/otlphttp/collector/trace/v1"
-	commonpb "go.opentelemetry.io/proto/otlphttp/common/v1"
-	resourcepb "go.opentelemetry.io/proto/otlphttp/resource/v1"
-	tracepb "go.opentelemetry.io/proto/otlphttp/trace/v1"
+	coltracepb "go.opentelemetry.io/proto/slim/otlp/collector/trace/v1"
+	commonpb "go.opentelemetry.io/proto/slim/otlp/common/v1"
+	resourcepb "go.opentelemetry.io/proto/slim/otlp/resource/v1"
+	tracepb "go.opentelemetry.io/proto/slim/otlp/trace/v1"
 )
 
 // ExportTraceServiceRequest corresponds to coltracepb.ExportTraceServiceRequest.