+++ exporters/otlp/otlptrace/otlptracehttp/v2/internal/transform/instrumentation.go
@@ -4,7 +4,7 @@
 package transform
 
 import (
-	commonpb "go.opentelemetry.io/proto/otlphttp/common/v1"
+	commonpb "go.opentelemetry.io/proto/slim/otlp/common/v1"
 
 	"go.opentelemetry.io/otel/sdk/instrumentation"
 )